+++ C18 Ex01 Ori 205665508 Amir 305707044/Ex01.FacebookAppWinformsUI/ICreateUIControl.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Windows.Forms;
+
+namespace Ex01.FacebookAppWinformsUI
+{
+    public interface ICreateUIControl
+    {
+        bool Invoked { get; }
+        void Start();
+        Control CreateUIControl();
+        Control Update();
+    }
+}
